hr_system/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
Jinja2==3.1.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
itsdangerous==2.2.0
aiosqlite==0.20.0
